Check that interface constructors and default class constructors are compatible. 1. Number of required parameters. 2. Names and order of named parameters. 3. [disabled] type warning for non-identical types. 4. [wait for Eric, not implemented] type parameters check.
R=zundel@google.com
BUG=
TEST=

Review URL: http://codereview.chromium.org//8786002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@2124 260f80e4-7a28-3924-810f-c04153c831b5

+++ dart/client/samples/total/src/DateTimeUtils.dart
@@ -26,7 +26,7 @@ class DateUtils {
   // Return a Date object corresponding to the given number of days after 1899-12-30.
   static Date getDateTime(double daysSince18991230) {
     double daysSinceEpoch = daysSince18991230 - DAYS_FROM_1900_TO_1970;
-    double millisSinceEpoch = daysSinceEpoch * MILLISECONDS_PER_DAY;
+    int millisSinceEpoch = (daysSinceEpoch * MILLISECONDS_PER_DAY).toInt();
     return new Date.fromEpoch(millisSinceEpoch, const TimeZone.utc());
   }
